+++ src/pages/home_page/home_page.jsx
@@ -1,7 +1,7 @@
 import InfiniteScroll from 'react-infinite-scroller';
-import useInfiniteMovieData from '../../hooks/use-infinite-movie-data';
-import PreviewList from './components/preview_list';
 import TopButton from '../../components/top_button';
+import useInfiniteMovieData from '../../hooks/use_infinite_movie_data';
+import PreviewList from './components/preview_list';
 
 const HomePage = () => {
 	const { data, fetchNextPage, isLoading, hasNextPage } = useInfiniteMovieData({
